Issue-52 Migration of JMAP test to Gatling 3
