+++ project/plugins.sbt
@@ -1,4 +1,3 @@
 logLevel := Level.Warn
 
-addSbtPlugin("io.gatling" % "gatling-sbt" % "2.2.0")
-addSbtPlugin("com.typesafe.sbteclipse" % "sbteclipse-plugin" % "4.0.0")
+addSbtPlugin("io.gatling" % "gatling-sbt" % "3.0.0")
