--- src/main/scala-2.11/org/apache/james/gatling/smtp/PreDef.scala
@@ -1,7 +1,0 @@
-package org.apache.james.gatling.smtp
-
-object PreDef {
-
-  implicit def smtpProtocolBuilder2SmtpProtocol(builder: SmtpProtocolBuilder): SmtpProtocol = builder.build()
-
-}
