+++ src/main/scala-2.12/org/apache/james/gatling/jmap/JmapAuthentication.scala
