+++ src/main/scala-2.12/org/apache/james/gatling/utils/JmapChecks.scala
