--- src/test/scala-2.11/org/apache/james/gatling/jmap/scenari/JmapAuthenticationScenario.scala
@@ -1,16 +1,0 @@
-package org.apache.james.gatling.jmap.scenari
-
-import io.gatling.core.Predef._
-import org.apache.james.gatling.control.UserCreator
-import org.apache.james.gatling.jmap.scenari.common.HttpSettings
-import org.apache.james.gatling.jmap.scenari.common.CommonSteps
-import org.apache.james.gatling.jmap.scenari.common.Configuration._
-
-class JmapAuthenticationScenario extends Simulation {
-
-  val scn = scenario("JmapAuthentication")
-    .exec(CommonSteps.authentication(new UserCreator(BaseJamesWebAdministrationUrl).createUsersWithInboxAndOutbox(UserCount)))
-
-  setUp(scn.inject(atOnceUsers(UserCount))).protocols(HttpSettings.httpProtocol)
-
-}
