--- src/test/scala-2.11/org/apache/james/gatling/jmap/scenari/common/HttpSettings.scala
@@ -1,13 +1,0 @@
-package org.apache.james.gatling.jmap.scenari.common
-
-import io.gatling.core.Predef._
-import io.gatling.http.Predef._
-
-object HttpSettings {
-
-  val httpProtocol = http
-    .baseURL(Configuration.BaseJmapUrl)
-    .acceptHeader("application/json")
-    .contentTypeHeader("application/json; charset=UTF-8")
-
-}
