+++ src/test/scala-2.12/org/apache/james/gatling/smtp/scenari/common/Configuration.scala
@@ -1,0 +1,10 @@
+package org.apache.james.gatling.smtp.scenari.common
+
+import org.apache.james.gatling.smtp.{SmtpActionBuilder, SmtpProtocol}
+
+object Configuration {
+
+  val smtp = SmtpProtocol.default
+
+  def smtp(requestName: String) = new SmtpActionBuilder(requestName, null, null)
+}
